Ticket #10 - Fixed some code.

+++ TravellingSalesman/Utils/CenteringConverter.cs
@@ -8,15 +8,15 @@ namespace TravellingSalesman.Utils
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            if (value is double && parameter is string)
+            if (value is int && parameter is string)
             {
-                double offset;
-                if (!double.TryParse((string)parameter, out offset))
+                int offset;
+                if (!int.TryParse((string)parameter, out offset))
                 {
                     offset = 0;
                 }
 
-                return (double)value - (offset / 2.0);
+                return (int)value - (offset / 2.0);
 
             }
             return value;
